requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
open-clip-torch>=2.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
